[Feat] 수정 모드 UI 구현 #27

+++ data/src/main/java/com/project200/data/api/ApiService.kt
@@ -21,6 +21,6 @@ interface ApiService {
     // 운동 기록 상세 조회
     @GET("v1/exercises/{recordId}")
     suspend fun getExerciseRecordDetail(
-        @Path("recordId") recordId: Int
+        @Path("recordId") recordId: Long
     ): BaseResponse<GetExerciseRecordData>
 }
